Dont build default.metallib when building ANGLE https://bugs.webkit.org/show_bug.cgi?id=259509 <rdar://112819533> Reviewed by Alex Christensen.
The autogenerated metal file was accidently included as a
build target in ANGLE's default shared library. It's previously
been compiled in a seperate project that builds to
"ANGLEMetalLib.metallib". After the last ANGLE roll,
it was building to both "default.metallib" and 'ANGLEMetalLib.metallib'

This change restores the previous behavior for this library.

* Source/ThirdParty/ANGLE/ANGLE.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/266307@main