[GLib] In new API, update hsts directories for TestWebsiteData. https://bugs.webkit.org/show_bug.cgi?id=259515
Reviewed by Michael Catanzaro.

As mentioned in 260511@main and 261018@main, A GLib-specific bug in the
defaultHSTSStorageDirectory is fixed, but only for the new API versions.
The HSTS database goes under the base cache directory if provided, but
if not provided it goes under the base data directory; leave it broken in
the original API versions to avoid leaking the HSTS database on disk.
Also, move it to a subdirectory in the new API version.

Updating the directories (to baseCacheDirectory for old API and
under subdirectory for new API, using code that we previously had before
259132@main) in prepopulateHstsData fixes the hsts test in
TestWebsiteData, which were marked as failing under a meta-gardening bug
in webkit.org/b/256557.

* Tools/TestWebKitAPI/Tests/WebKitGLib/TestWebsiteData.cpp:
(prepopulateHstsData): add HSTS subdir for new API for hsts tests,
this was code we previously used prior to 259132@main.
* Tools/TestWebKitAPI/glib/TestExpectations.json: Remove WPE failure
for hsts test.

Canonical link: https://commits.webkit.org/266326@main