Specialize appendChild for cloneNodeInternal https://bugs.webkit.org/show_bug.cgi?id=259542
Reviewed by NOBODY (OOPS!).

Merge the contents of appendChild into cloneNodeInternal. This patch also introduces ChildChange::Source::Clone to
distinguish cloning operation in childrenChanged function calls.

* Source/WebCore/dom/ContainerNode.cpp:
(WebCore::ContainerNode::removeNodeWithScriptAssertion): Assert that source is either Source::API or Source::Parser.
(WebCore::makeChildChangeForCloneInsertion): Added.
(WebCore::ContainerNode::cloneChildNodes): Merged appendChild into this function.

* Source/WebCore/dom/ContainerNode.h:

* Source/WebCore/html/HTMLDetailsElement.cpp:
(WebCore::HTMLDetailsElement::create): Create an event allowed scope on the newly created element as we're about to add
a shadow root to it.

* Source/WebCore/html/HTMLOptionElement.cpp:
(WebCore::HTMLOptionElement::childrenChanged): Don't notify select element if we're in the middle of cloning operation.
Without this change, we would hit a debug assertion in HTMLSelectElement::listItems() because cloning operation now
calls childrenChanged on an option element before the select element.

* Source/WebCore/html/HTMLTextAreaElement.cpp:
(WebCore::HTMLTextAreaElement::create): Create an event allowed scope on the newly created element as we're about to
add a shadow root to it.