[RTCVideoEncoderH264 initWithCodecInfo] should cope with invalid profile level id https://bugs.webkit.org/show_bug.cgi?id=259592 rdar://113008968
Reviewed by Eric Carlson.

With WebCodecs, we allow JS to pass any profile level id to RTCVideoEncoderH264.
For now, we are defaulting to a default profile level id if none is matching.
A future patch should probably make sure that we return an error so that JS knows the configuration is not unsupported.

* Source/ThirdParty/libwebrtc/Source/webrtc/api/video_codecs/h264_profile_level_id.cc:

Canonical link: https://commits.webkit.org/266392@main