Canvas does not show results of CanvasRenderingContext2D.putImageData until forced re-render. https://bugs.webkit.org/show_bug.cgi?id=259584 rdar://112901862
Reviewed by Tim Nguyen.

The commit in 263694@main caused a regression when calculating the dirty region
of a putImageDataUpdate. It accidentally removed a reference marker on a parameter.

* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::computeImageDataRect): Make sure the destRect parameter is a reference because it is updated.
(WebCore::CanvasRenderingContext2DBase::putImageData): Leave a fixme!

Canonical link: https://commits.webkit.org/266397@main