AX: display:contents elements are sometimes missing their children https://bugs.webkit.org/show_bug.cgi?id=259608 rdar://problem/113044333
Reviewed by Chris Fleizach.

This happened because AccessibilityObject::insertChild detects and bails when an object is trying to insert a child
that belongs to a display:contents object that is not `this`. But if the correct parent of that child does not have
it’s dirty-children bit set, we never actually insert the child, resulting in it being missing from the accessibility
tree. With this patch, we set that bit, ensuring the accessibility tree is updated correctly.

* LayoutTests/accessibility/display-contents/table-dynamic-expected.txt: Added.
* LayoutTests/accessibility/display-contents/table-dynamic.html: Added.
* LayoutTests/platform/ios/accessibility/display-contents/table-dynamic-expected.txt: Added.
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::insertChild):

Canonical link: https://commits.webkit.org/266407@main