[Xcode] Add WK_PRIVATE_SDK_DIR for determining which SDK has private frameworks https://bugs.webkit.org/show_bug.cgi?id=259541 rdar://112945796
Reviewed by Alexey Proskuryakov.

Defines a build setting for whichever SDK is expected to contain private
frameworks. Internally and on macOS, this is always the base SDK, but it
is one of the additions SDKs in open source embedded builds.

Renames the canonical names of each sparse SDK to "WebKitSDKAdditions".
Since we should never be using more than one of our sparse SDKs at once,
this is fine, and it makes it possible to resolve the sparse SDK's
directory with $(SDK_DIR_WebKitSDKAdditions).

Finally, update all search paths used to find and link
against private frameworks to refer to WK_PRIVATE_SDK_DIR instead of
SDK_DIR. This allows us to remove the search path declarations in every
SDKSettings.plist; the same variables that put private frameworks on the
search path internally are now used in open source.

* Configurations/SDKAdditions.xcconfig:
* Source/JavaScriptCore/Configurations/Base.xcconfig:
* Source/WTF/Configurations/Base.xcconfig:
* Source/WTF/Configurations/WTF.xcconfig:
* Source/WebCore/Configurations/WebCore.xcconfig:
* Source/WebCore/PAL/Configurations/PAL.xcconfig:
* Source/WebCore/PAL/ThirdParty/libavif/Configurations/libavif.xcconfig:
* Source/WebCore/PAL/ThirdParty/libavif/ThirdParty/dav1d/Configurations/dav1d.xcconfig:
* Source/WebGPU/Configurations/WGSL.xcconfig:
* Source/WebGPU/Configurations/WebGPU.xcconfig:
* Source/WebGPU/Configurations/wgslc.xcconfig:
* Source/WebInspectorUI/Configurations/Base.xcconfig:
* Source/WebKit/Configurations/BaseTarget.xcconfig:
* Source/WebKitLegacy/mac/Configurations/WebKitLegacy.xcconfig:
* Tools/DumpRenderTree/mac/Configurations/Base.xcconfig:
* Tools/TestWebKitAPI/Configurations/Base.xcconfig:
* WebKitLibraries/SDKs/appletvos16.0-additions.sdk/SDKSettings.plist:
* WebKitLibraries/SDKs/iphoneos16.0-additions.sdk/SDKSettings.plist:
* WebKitLibraries/SDKs/macosx12.0-additions.sdk/SDKSettings.plist:
* WebKitLibraries/SDKs/macosx12.0.internal-additions.sdk/SDKSettings.plist:
  In internal builds where we inject additional frameworks (e.g.
  AppleJPEGXL.framework), we don't want to override where the compiler
  searches for private frameworks; only make our additional frameworks
  findable. Keep the search path listed in the SDK to achieve this, and
  add a comment explaining why.
* WebKitLibraries/SDKs/macosx13.0-additions.sdk/SDKSettings.plist:
* WebKitLibraries/SDKs/macosx13.0.internal-additions.sdk/SDKSettings.plist: Ditto internal change.
* WebKitLibraries/SDKs/macosx14.0-additions.sdk/SDKSettings.plist:
* WebKitLibraries/SDKs/watchos9.0-additions.sdk/SDKSettings.plist:

Canonical link: https://commits.webkit.org/266463@main