REGRESSION(263118@main): [GTK] Web inspector does not paint reliably https://bugs.webkit.org/show_bug.cgi?id=259449
Reviewed by Michael Catanzaro.

This happens with the inspector because the inspector view doesn't
follow the hardware accelerated settings, and always runs in
non-accelerated compositing mode. This is actually a bug in
non-accelerated compositing mode implementation, it happens with any web
view when disabling hardware acceleration. The problem is that when the
view is no longer in the active window (it loses the focus) we discard
the backing store after 2 seconds and we don't notify the web process
nor ask for an update when a new paint is requested (when the web view
is redraw), as we did before 263118@main. After the resize
simplification we can simplify this part too, and just add a message to
request an update when the web view is redrawn after the backing store
has been discarded. I'm also increasing the delay to discard the backing
store, because with 2 seconds we usually end up creating a new one after
losing the focus due to the repaint requested by the overlay scrollbars
fade out animation. Using 10 seconds we ensure we always discard the
backing store after the scrollbar animation finished and no more repaints
are expected.

* Source/WebKit/UIProcess/CoordinatedGraphics/DrawingAreaProxyCoordinatedGraphics.cpp:
(WebKit::DrawingAreaProxyCoordinatedGraphics::paint):
(WebKit::DrawingAreaProxyCoordinatedGraphics::discardBackingStoreSoon):
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/DrawingAreaCoordinatedGraphics.cpp:
(WebKit::DrawingAreaCoordinatedGraphics::forceUpdate):
* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/DrawingAreaCoordinatedGraphics.h:
* Source/WebKit/WebProcess/WebPage/DrawingArea.h:
(WebKit::DrawingArea::forceUpdate):
* Source/WebKit/WebProcess/WebPage/DrawingArea.messages.in:

Canonical link: https://commits.webkit.org/266511@main