[GTK][WPE] API test WebKitWebsiteData/appcache consistently passing https://bugs.webkit.org/show_bug.cgi?id=188117
Unreviewed test gardening.

This test was marked as flaky in `gtk-linux-64-release/build` (bot 41)
in b/188117 in 2018; this test is now passing in that bot, as well as
on GTK-Linux-Release-Tests bots and on recent passes of WPE Test release
bots (see also `wk-bot-digest`, as this test passes in the last few
thousand reports for both WPE and GTK; WPE bots have been exiting early
for the last few weeks which is being worked on separately).

* Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/266522@main